feat(admin) add role and user data for  get admin endpoints

+++ app/routers/admin.py
@@ -1,7 +1,7 @@
 # app/routers/admin.py
 
 from uuid import UUID
-from fastapi import APIRouter, Depends, HTTPException, status, Security
+from fastapi import APIRouter, Depends, HTTPException, status
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.future import select
 from sqlalchemy.orm import selectinload
@@ -11,25 +11,24 @@ from app.utils import (
     get_current_admin,
     get_superadmin,
     logger,
-    get_moderator,
-    get_support_admin,
-    get_content_manager,
-    get_by_email,
+    get_user_by_email,
     create_user,
     get_role_by_name,
     get_admin_by_id,
-    get_role_by_id
+    get_role_by_id,
+    get_user_by_id
 )
-from app.schemas.admin import RoleCreate, RoleUpdate, PermissionUpdate, AdminCreate, AdminUpdate
+from app.schemas.admin import AdminResponse, AdminCreate, AdminUpdate
 
 router = APIRouter(prefix="/admin", tags=["admin"])
 
 # ------------------------------ Role Management Endpoints ------------------------------
 
+
 @router.get("/roles")
 async def list_roles(
     db: AsyncSession = Depends(get_db),
-    current_admin: User = Depends(get_current_admin)  # Any admin can list roles
+    current_admin: User = Depends(get_current_admin),  # Any admin can list roles
 ):
     """
     List all roles.
@@ -41,7 +40,7 @@ async def list_roles(
         logger.error(f"Error listing roles: {e}")
         raise HTTPException(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Internal server error"
+            detail="Internal server error",
         )
 
 
@@ -49,7 +48,7 @@ async def list_roles(
 async def get_role(
     role_id: UUID,
     db: AsyncSession = Depends(get_db),
-    current_admin: User = Depends(get_current_admin)  # Any admin can view roles
+    current_admin: User = Depends(get_current_admin),  # Any admin can view roles
 ):
     """
     Get details of a specific role.
@@ -59,37 +58,38 @@ async def get_role(
         role = role.scalars().first()
         if not role:
             raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="Role not found"
+                status_code=status.HTTP_404_NOT_FOUND, detail="Role not found"
             )
         return role
     except Exception as e:
         logger.error(f"Error fetching role: {e}")
         raise HTTPException(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Internal server error"
+            detail="Internal server error",
         )
 
+
 # ------------------------------ Admin User Management Endpoints ------------------------------
+
 
 @router.post("/", status_code=status.HTTP_201_CREATED)
 async def create_admin(
     admin_data: AdminCreate,
     db: AsyncSession = Depends(get_db),
-    current_admin: User = Depends(get_superadmin)  # Superadmin can create admins
+    current_admin: User = Depends(get_superadmin),  # Superadmin can create admins
 ):
     """
     Create a new admin.
     """
     try:
-        existing_user = await get_by_email(db, admin_data.email)
+        existing_user = await get_user_by_email(db, admin_data.email)
         if existing_user:
             raise HTTPException(
                 status_code=status.HTTP_400_BAD_REQUEST,
                 detail="User with this email already exists",
             )
-        
-        new_user = await create_user(db, admin_data,is_admin=True)
+
+        new_user = await create_user(db, admin_data, is_admin=True)
 
         role = await get_role_by_name(db, name=admin_data.admin_sub_role)
         if not role:
@@ -102,7 +102,9 @@ async def create_admin(
         admin = Admin(id=new_user.id, role_id=role.id)
         db.add(admin)
         await db.commit()
-        logger.info(f"Admin created for user ID '{admin.id}' by admin '{current_admin.email}'.")
+        logger.info(
+            f"Admin created for user ID '{admin.id}' by admin '{current_admin.email}'."
+        )
         return {"message": "Admin created successfully", "admin_id": admin.id}
     except HTTPException:
         raise
@@ -110,27 +112,41 @@ async def create_admin(
         logger.error(f"Error creating admin: {e}")
         raise HTTPException(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Internal server error"
+            detail="Internal server error",
         )
 
 
-@router.get("/")
+@router.get("/", response_model=list[AdminResponse])
 async def list_admins(
     db: AsyncSession = Depends(get_db),
-    current_admin: User = Depends(get_current_admin)  # Any admin can list admins
+    current_admin: User = Depends(get_current_admin),  # Any admin can list admins
 ):
     """
     List all admins.
     """
     try:
-        admins = await db.execute(select(Admin))
-        return admins.scalars().all()
+        admins = await db.execute(
+            select(Admin).options(selectinload(Admin.role), selectinload(Admin.user))
+        )
+        result = admins.scalars().all()
+        return [
+            {
+                "created_at": admin.created_at,
+                "updated_at": admin.updated_at,
+                "role_id": admin.role_id,
+                "role_name": admin.role.name,
+                "id": admin.id,
+                "full_name": admin.user.full_name,
+            }
+            for admin in result
+        ]
     except Exception as e:
         logger.error(f"Error listing admins: {e}")
         raise HTTPException(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Internal server error"
+            detail="Internal server error",
         )
+
 
 @router.get("/sub-role")
 async def get_admin_sub_role(
@@ -148,11 +164,11 @@ async def get_admin_sub_role(
     return {"admin_sub_role": admin.role.name}
 
 
-@router.get("/{admin_id}")
+@router.get("/{admin_id}", response_model=AdminResponse)
 async def get_admin(
     admin_id: UUID,
     db: AsyncSession = Depends(get_db),
-    current_admin: User = Depends(get_current_admin)  # Any admin can view admins
+    current_admin: User = Depends(get_current_admin),  # Any admin can view admins
 ):
     """
     Get details of a specific admin.
@@ -161,15 +177,21 @@ async def get_admin(
         admin = await get_admin_by_id(db, admin_id)
         if not admin:
             raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="Admin not found"
+                status_code=status.HTTP_404_NOT_FOUND, detail="Admin not found"
             )
-        return admin
+        return {
+            "created_at": admin.created_at,
+            "updated_at": admin.updated_at,
+            "role_id": admin.role_id,
+            "role_name": admin.role.name,
+            "id": admin.id,
+            "full_name": admin.user.full_name,
+        }
     except Exception as e:
         logger.error(f"Error fetching admin: {e}")
         raise HTTPException(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Internal server error"
+            detail="Internal server error",
         )
 
 
@@ -178,7 +200,7 @@ async def update_admin(
     admin_id: UUID,
     admin_data: AdminUpdate,
     db: AsyncSession = Depends(get_db),
-    current_admin: User = Depends(get_superadmin)  # Superadmin can update admins
+    current_admin: User = Depends(get_superadmin),  # Superadmin can update admins
 ):
     """
     Update an admin.
@@ -187,8 +209,7 @@ async def update_admin(
         admin = await get_admin_by_id(db, admin_id)
         if not admin:
             raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="Admin not found"
+                status_code=status.HTTP_404_NOT_FOUND, detail="Admin not found"
             )
 
         # Update admin fields
@@ -202,7 +223,7 @@ async def update_admin(
         logger.error(f"Error updating admin: {e}")
         raise HTTPException(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Internal server error"
+            detail="Internal server error",
         )
 
 
@@ -210,7 +231,7 @@ async def update_admin(
 async def delete_admin(
     admin_id: UUID,
     db: AsyncSession = Depends(get_db),
-    current_admin: User = Depends(get_superadmin)  # Superadmin can delete admins
+    current_admin: User = Depends(get_superadmin),  # Superadmin can delete admins
 ):
     """
     Delete an admin.
@@ -219,16 +240,22 @@ async def delete_admin(
         admin = await get_admin_by_id(db, admin_id)
         if not admin:
             raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail="Admin not found"
+                status_code=status.HTTP_404_NOT_FOUND, detail="Admin not found"
+            )
+        
+        user = await get_user_by_id(db, admin.id)
+        if not user:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND, detail="User not found"
             )
 
         await db.delete(admin)
+        await db.delete(user)
         await db.commit()
         logger.info(f"Admin '{admin.id}' deleted by admin '{current_admin.email}'.")
     except Exception as e:
         logger.error(f"Error deleting admin: {e}")
         raise HTTPException(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Internal server error"
+            detail="Internal server error",
         )
